coffee_guide/pyproject.toml
[tool.poetry]
name = "coffee-guide-backend"
version = "0.1.0"
description = "Веб-сервис Кофейный гид"
authors = ["-"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djoser = "^2.2.2"
pillow = "^10.1.0"
requests = "^2.31.0"
drf-spectacular = "^0.27.0"
drf-spectacular-sidecar = "^2023.12.1"
social-auth-app-django = "^5.4.0"
django-filter = "^23.5"
dadata = "^21.10.1"
psycopg2-binary = "^2.9.9"





[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
flake8 = "^6.1.0"
black = {version = "^23.11.0", allow-prereleases = true}
mypy = "^1.7.1"
isort = "^5.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
